LPS-104616 Move addFragment to thunks
